Moved local topic client into local_proxy folder
