+++ ydb/core/kqp/federated_query/kqp_federated_query_helpers.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <ydb/core/base/appdata.h>
-#include <ydb/core/kqp/federated_query/local_pq_client/local_topic_client_settings.h>
+#include <ydb/core/local_proxy/local_pq_client/local_topic_client_settings.h>
 #include <ydb/library/logger/actor.h>
 #include <ydb/library/yql/providers/common/db_id_async_resolver/db_async_resolver.h>
 #include <ydb/library/yql/providers/common/db_id_async_resolver/mdb_endpoint_generator.h>
