+++ ydb/core/local_proxy/ya.make
@@ -1,0 +1,3 @@
+RECURSE(
+    local_pq_client
+)
