+++ ydb/core/ya.make
@@ -36,6 +36,7 @@ RECURSE(
     kqp
     load_test
     local_pgwire
+    local_proxy
     log_backend
     metering
     mind
